igvm: Implement launch of OVMF firmware from IGVM file
When using an IGVM file the OVMF firmware will be populated into the
guest using pages described by IGVM and not using firmware volumes and
fw_cfg.

This patch uses the current configuration (IGVM or fw_cfg) to determine
the firmware range and validate the pages accordingly before launching
the firmware.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>