svsm: Allow the firmware metadata location to be specified via config
When using fw_cfg, the OVMF metadata is located at a hardcoded location
of 1 page below 4GB. When using IGVM for configuration it is possible
for the IGVM file to specify a different location for the firmware
itself so it makes sense to also specify a parameter to describe the
metdata page location. This commit adds a new IGVM parameter field and
uses this to determine the OVMF metadata location.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>