igvmbld: Populate Checksum in IGVM_FIXED_HEADER
The checksum field in the fixed header should be calculated from the
fixed and variable headers in the generated file. This is currently just
set to 0. The rust IGVM library validates the checksum field when
reading the file so files generated with the current version of the
builder cannot be loaded when using the rust library.

This commit calculates the checksum and populates it in the generated
file.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>