utils: Add a StringRingBuffer implementation
The log buffer is implemented using a ring buffer that contains the
characters within log entries. This patch introduces a generic
StringRingBuffer that can be used for safely managing the contents of
strings within a ring buffer that is suitable for use with the log
buffer. This separates the ring buffer implementation from the log and
allows implementation of ring buffer specific unit testing, which is
included within this patch.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>