Change SEV_STATUS check from not-supported to supported
Since the set of defined SEV features can change over time, it is not
possible for the code to stay current with the set of features that
should not be supported.  Instead, it is safer for the code to enumerate
the set of features that can be supported, with the assumption that any
unknown feature must be rejected.

BTB isolation and debug register isolation offer security and do not
have any impact to software functionality so they can safely be accepted
if present.  Other features may require explicit software support.

Signed-off-by: Jon Lange <jlange@microsoft.com>