igvmbld: Allow firmware to be built into IGVM file
This commit adds parameters to the IGVM builder command line that allow
a firmware binary file to be specified along with either the firmware
base address or top address. The firmware is then populated into the
IGVM file at the specified address.

If the firmware contains metadata in OVMF format then this is parsed and
the relevant fields of the IGVM parameter block that is passed to the
guest.

This allows OVMF (and potentially other) firmware to be embedded in the
IGVM file and launched from the SVSM.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>