svsm: Use fw metadata from IGVM config instead of from OVMF
When using an IGVM based configuration, currently the firmware metadata
can be provided either in the IGVM configuration itself or can be parsed
from the OVMF image. However, there is no reason for SVSM to parse the
OVMF metadata as this can be done by the IGVM builder and populated into
the IGVM configuration.

This patch removes the logic to parse OVMF metadata when using IGVM
configuration and applies the required extra fields from the IGMV
configuration.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>