igvm: Refactor firmware configuration in IGVM parameter block
Perform a refactor of the firmware configuration fields within the IGVM
parameter block and add the additional fields required by OVMF firmware.

This neatly separates the firmware metadata from the remaineder of the
IGVM parameters and makes it clearer which fields relate to the firmware
rather than the SVSM itself.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>