SVSM/cpu: Ensure that guest VMPLs do not enable restricted injection
Guest VMPLs will not enable exactly the same SEV features as SVSM VMPLs.
In particular, guest VMPLs will not use restricted injection even if the
SVSM does, and in the future, guest VMPLs may need to enable (or
disable) other features such as alternate injection, REFLECT_VC, or
vTOM.  This change refactors the determination of SEV features for guest
VMPLs so that it is possible to change features in a robust way.

Signed-off-by: Jon Lange <jlange@microsoft.com>