Move shared CPU state into a separate structure
In order to preserve the Rust guarantees of mutually exclusive mutable
and immutable references, it must never be possible for a PerCpu object
to be accessible across threads, even to access fields that are designed
to be thread-safe.  This change moves globally shared per-CPU fields
into a separate structure that supports only immutable references, and
fields contained within are synchronized using appropriate locking
primitives that permit safe interior mutability.

Signed-off-by: Jon Lange <jlange@microsoft.com>