igvmbld: Extend verbose mode to show guest memory and igvm params
In order to help validate and debug issues around allocation of guest
physical address ranges, and to ensure it matches the expectations of
coconut SVSM, it is useful to see what is generated by the IGVM file
builder.

This commit extends the verbose output from the tool to show the purpose
and the range for each block of guest physical memory defined by the
generated IGVM file as well as giving a detailed summary of the contents
of the IGVM parameter block.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>