svsm: Invalidate IGVM parameter area after stage2
When booting using an IGVM configuration, some of the prevalidated pages
are populated outside the first 640K of low memory. These need to be
invalidated prior to booting the firmware.

This patch passes the location and size of the IGVM parameter block
which consitutes all of the prevalidated page outside the 640K range
from stage2 to the kernel and invalidates them at the correct time.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>