igvmbld: Fix merge error that caused overlapping memory regions
The change made in commit 571db22fbaba413a9299b14aafa294e251ed646c that
reduces the memory range for stage2 was lost in a subsequent commit
causing two memory ranges to overlap in the IGVM file. This is
subsequently rejected by the QEMU builder causing the guest to fail to
startup.

This commit fixes the merge error.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>