Drop local references to VMSA while processing a request
The VMSA reference must be dropped during request processing because the
VMSA reference is tied to the life of the VMSA spinlock.  If the
reference were not to be dropped, then the acquisition of the VMSA
inside the request processor would deadlock as it contends for the lock
held by the request loop.  Dropping the lock from the request loop
prevents this deadlock.

Signed-off-by: Jon Lange <jlange@microsoft.com>