igvmbld: Zero stage2 stack page before populating it
The stage2 stack page memory is allocated with a call to
construct_mem_data_object(). This function zeros any padding outside of
the requested data size but does not zero the memory where the data
resides.

The stage2 stack is allocated as a full page, meaning the stack is not
initialised to zero. This results in uninitialised data being present
below the populated stack frame and in the _reserved part of the stack.

This commit zeroes the entire stack page to ensure it is initialised.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>