SVSM/cpu/gdt: Fix boot issue around TSS loading in release builds
In release builds the compiler did not recognize that the assembly
between set_tss_entry() and clear_tss_entry() did actually consume the
data written by set_tss_entry(). This make the compiler optimize away
the set_tss_entry() call entirely in release builds, making the SVSM
fail to boot.

Use volatile writes to update the GDT for TSS entries to force the
compiler not to optimize the writes away.

Signed-off-by: Joerg Roedel <jroedel@suse.de>