Copy IGVM builder to bin directory
To support external workflows that are not incorporated into the
makefile, publish the IGVM file builder to the bin directory so it can
be consumed by those workflows.

Also, this changes the makefile so that creation of the bin directory is
coupled with specific actions that target the bin directory, instead of
an unrelated build target.

Signed-off-by: Jon Lange <jlange@microsoft.com>