Fix incremental build
The binary targets generated by cargo must be listed as .PHONY targets
to ensure that make knows to invoke cargo to force the targets to come
up to date.  Without this, executing "make" will not reevaluate the Rust
sources to determine whether rebuilds of the kernel or stage2 are
necessary.

Signed-off-by: Jon Lange <jlange@microsoft.com>