Makefile: Move all generated binary targets to bin dir
Besides being more consistent, this improves the process of performing a
clean make ("make clean") because all generated targets are in one place
and can be removed more consistently.

The "make clean" target continues to refer to the old files in the stage1
directory so that executing "make clean" after picking up this change
will correctly remove the old files.

Signed-off-by: Jon Lange <jlange@microsoft.com>