igvmbuilder: use Vec::extend_from_slice() when building IGVM directives
Vec::append() moves elements from one Vec to another, leaving the
first one empty. However, we are doing so from a clone of
self.firmware, so effectively we are cloning a Vec and draining it in
place. Instead, simpy use Vec::extend_from_slice, which copies
elements from a reference directly.

Signed-off-by: Carlos López <carlos.lopez@suse.com>