igvmbuilder/cpuid: avoid unsafe block
SnpCpuidPage derives the AsBytes trait, so simply use it to get the
type represented as a u8 slice. Since we are converting the buffer
into a Vec later, do it upfront from the returned slice and fill with
zeroes up to the desired 4k size. This avoids an unnecessary initial
memset().

We can also assert the type's size matches what we expect at compile
time, so remove the runtime check.

Signed-off-by: Carlos López <carlos.lopez@suse.com>