test-in-svsm: Switch to using IGVM to package the test-in-svsm binary
The documentation has been updated to use IGVM when packaging and
launching the COCONUT-SVSM binary in a QEMU SEV-SNP supported guest in
preparation for removing support for the previous QEMU SVSM
modifications. Therefore, the test-in-svsm Makefile target should also
be updated to use IGVM to launch the guest that runs the tests.

This patch builds adds a new Makefile target to build an IGVM file that
contains the svsm-test binary and makes use of scripts/launch_guest.sh
to launch the SVSM test binary using QEMU.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>