gdbstub: store a pointer in GdbStubTarget
We were storing in GdbStubTarget a usize that acted as a pointer via
"as" casting. Simply store the pointer and add the necessary Send and
Sync implementations that we were working our way around.

Add two methods to perform the unsafe conversion from the pointer to
the exception context to a safe reference and properly document their
safety.

Signed-off-by: Carlos López <carlos.lopez@suse.com>