kernel/cpu/idt: handle nested #VC
Now that #VC handler runs on an IST stack, nested #VC can be handled
on the same stack than the parent handler, possibly overwriting the
parent's stack content.

We need to handle nested #VC differently. We can detect if #VC is
nested by checking the value of RSP pushed on the stack in the early
exception handler. We can also detect #VC raised from user-mode.
For both cases, we relocate the stack to the current task's stack that
is safe to use.

Finally, we need to copy the pushed registers to the new stack address,
and call the regular #VC handler like if #VC handling was not IST.

Signed-off-by: Thomas Leroy <thomas.leroy@suse.com>