stage2: Record the SVSM platform indicator in the stage2 stack
Recording the platform type in the stage 2 stack is necesary to enable
stack 2 (and later, the kernel) to know what platform configuration it
is running under so it knows which operations it should attempt to
perform as required by that target architecure.  The IGVM file embeds a
separate copy of the stage 2 stack per platform, and stage 1 encodes SNP
as that is the only supported non-IGVM platform.

Signed-off-by: Jon Lange <jlange@microsoft.com>