platform: Define an abstraction to describe the base platform
Because the SVSM will need to run on multiple platform types, it is
useful to aggregate platform-specific functionality into an object that
can be implemented separately for each possible target platform.  This
change defines the basic abstraction.  It is implemented as a dynamic
dispatch type to avoid the proliferation of generic implementations
among the many callers that will exist across the code base that require
platform-specific functionality.  Initially, all platform types will be
compiled into the binary, though in the future, it will be possible to
define compile-time configurations that limit the set of platform
implementations that will be included.

Signed-off-by: Jon Lange <jlange@microsoft.com>