igvmbuilder/vmsa: Ensure VMSA matches KVM requirements
When using the generated IGVM file in QEMU, the VMSA described by the
IGVM file must have specific values in certain fields in order to ensure
the initial VMSA that is presented to the guest matches that described
in the IGVM file. This is because the host kernel KVM will modify
certain registers/states before measuring the VMSA.

This patch ensures the IGVM VMSA requires no modification by KVM. This
allows the IGVM file to be measured, resulting in a measurement that
matches the SEV attestation report.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>