makefile: Run igvmmeasure after building IGVM files
This commit invokes igvmmeasure for each IGVM file that is built by the
makefile. This results in the launch measurement of each IGVM file being
output to the console for information. The developer can manually verify
this measurement by comparing it with the actual launch measurement
which is displayed in the SVSM log.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>