Prevent use of ImmutAfterInit in the multithreaded environment
The initialization of `ImmutAfterInitCell` objects (and, by extension,
`ImmutAfterInitRef` objects) is designed specifically only to safe when
the system is still in a single-threaded environment, because the
intialization checks are not thread-safe.  This change ensures that no
attempt is made to initialize or reinitialize any of these objects once
the system has started multiple processors.  These checks are
debug-only, like the checks on initialization state.

Signed-off-by: Jon Lange <jlange@microsoft.com>