Fix Hyper-V IGVM without firmware
In the Hyper-V configuration, the IGVM builder places the kernel image
just above the firmware image.  If there is no firmware image, this
results in the kernel image being placed at address zero.  This overlaps
with where stage 2 is loaded, and that results in corruption of the
kernel image and/or the IGVM parameter block depending on the extent of
the collision.  Moving the kernel to a minimum address of 1 MB ensures
that there is no collision.

Signed-off-by: Jon Lange <jlange@microsoft.com>