mm/pagetable: fix PageTableRef memory leaks
PageTableRefs are most of the times (always except during early
initialization) placed on newly allocated pages. However, when the
type is dropped the allocated page is never returned to the allocator.
This can become a problem in the long run if enough tasks are created,
as each allocates their own page table.

To fix this and stay compatible with non-allocated PageTableRefs, add
a new field that indicates whether the inner pointer was obtained from
the page allocator.

Signed-off-by: Carlos López <carlos.lopez@suse.com>