igvmbuilder: Add option to determine how to handle zero pages
The IGVM specification here defines how to handle pages that contain
only zeros here:
https://docs.rs/igvm_defs/0.1.6/igvm_defs/struct.IGVM_VHS_PAGE_DATA.html#fields

This states that if a platform supports a native page type for zero
pages then it should be used, such as is the case for SEV-SNP. The
IGVM measure tool measures according to the specification. However,
existing implementations of IGVM loaders instead populate a page
of zeroes and measure this using the normal page type.

This commit adds an option that allows the behaviour of the measure tool
to be defined when handling zero pages.

Note that it is expected that the IGVM specification will be modified to
change the expected behaviour so this option defaults to using native
pages for zeroes.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>