svsm/requests: introduce the SvsmCaa type
In the SVSM protocol, requests are serviced by reading and writing to
a structure shared with the guest. The fields of this structure fit
into a u64, so we were using that up until now. In order to improve
readability and future mistakes, add a proper type for this structure.

Signed-off-by: Carlos López <carlos.lopez@suse.com>