svsm: support SVSM discovery via CPUID
The SVSM spec states that post-boot elements may require discovering
the presence of the SVSM, and can do so by reading bit 28 in EAX for
CPUID(EAX=0x8000001F).

To do so, set this bit when initializing the CPUID page for the SVSM.
We do not need to do this for stage2 as post-boot elements will only
come only come online after the main SVSM kernel has started, at which
time the stage2 CPUID will no longer be in use.

Signed-off-by: Carlos López <carlos.lopez@suse.com>