kernel/cpu: Move per-CPU stack bounds into unsafe area
The panic handler needs to be able to access stack bounds, so it must
not require any references to the current per-CPU object via
`this_cpu()` or `this_cpu_mut()`.  The stack bounds are updated from
unsafe code to reinforce the fact that they are not protected by borrow
checks.

Signed-off-by: Jon Lange <jlange@microsoft.com>