kernel/cpu: move cpu online state to shared CPU structure
The contents of `PerCpu` can only ever be visible on the current CPU in
order to satsify borrow requirements.  However, the online state of the
CPU is consumed across processors.  This change moves the `online` field
to `PerCpuShared` so it can correctly be viewed across processors.

Additionally, `set_online` is changed to use release semantics to
correctly match the acquire semantics expected by `is_online`.

Signed-off-by: Jon Lange <jlange@microsoft.com>