igvmbuilder: Add option to provide guest policy
The guest policy can be included in the initialization header section in
the IGVM file. The policy affects the security level of the guest so
forms part of the launch measurement on the platform.

This adds a new parameter to igvmbuilder that allows the guest policy to
be specified as a hex string which then creates a new entry in the
initialization section of the IGVM file.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>