igvmmeasure: Refactor command line to introduce subcommands
The igvmmeasure tool is currently only used to display the measurement
of an IGVM file on the console. This is going to be expanded to include
the capability to sign the igvm file with the calculated measurement and
signatures.

In preparation for this, this commit introduces the 'measure' subcommand
which provides exactly the same functionality as currently provided but
will allow for new subcommands to be added in future commits.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>