igvmmeasure: Add capability to sign IGVM files
This extends the igvmmeasure tool with a new command, 'sign' which
allows a signed launch digest measurement to be added to a copy
of the input IGVM file in the form of an SnpIdBlock directive.

The igvmmeasure 'sign' command signs the digest using a private key
provided on the command line. The public part of this key can optionally
be signed with an author key if provided on the command line.

The resulting signed IGVM file can be provided to any IGVM loader that
supports SEV-SNP ID blocks.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>