kernel/protocols: Exclude parallel pvalidate and create_vcpu calls
Introduce a read-write lock to prevent a race condition around
parallel pvalidate and create_vcpu calls. The attack could be used to
give a guest OS access to a non-zeroed page, by using the error-path
of the create_vcpu call. The attack could lead to SVSM data exposure
to the guest OS.

Prevent that by taking the lock for read in the pvalidate path and for
write in the create_vcpu path.

Signed-off-by: Joerg Roedel <jroedel@suse.de>