mm/alloc: acquire ROOT_MEM lock for the whole function in layout_from_ptr()
In layout_from_ptr() we are acquiring the root memory lock twice back
to back. Simply acquire it once and hold it for the whole function -
that is the only way we can guarantee that, if the page we are
looking at is a slab page, the contained pointer will be valid when
we dereference it.

Signed-off-by: Carlos López <carlos.lopez@suse.com>