sev/ghcb: introduce GHCB field getters and setters
When accessing GHCB fields that were written to by the host, we must
check that they are set as valid in the valid bitmap. Conversely, when
setting fields of the GHCB to communicate with the host, we must set
them as valid in the valid bitmap. To do so, we must get the offset of
such field in the GHCB and translate it into an index into the valid
bitmap.

Instead of hardcoding GHCB field offsets, implement getters and
setters for each field, which will automatically get or set the
appropriate value from the valid bitmap based on that field's offset.
This was not done before because there was no way of getting a field's
offset at compile time before Rust 1.77, which introduced the
core::mem::offset_of!() macro.

Signed-off-by: Carlos López <carlos.lopez@suse.com>