mm/alloc: store item size instead of slab pointer in page info
The allocator stores information about each allocated page separately.
For slab pages, we store a pointer to such slab. We only use this to
be able to, from an allocation, obtain the size of the items in the
corresponding slab. Thus, it makes more sense to store the slab item
size instead. This reduces the amount of unsafe code, and removes a
level of indirection in layout_from_ptr() that we had when
dereferencing the stored slab pointer.

Signed-off-by: Carlos López <carlos.lopez@suse.com>