console: initialize WRITER from init_console()
The initialization of the console (`WRITER`) and marking it as
initialized (`init_console()`) are currently split from the
perspective of the caller. Merge those two steps into a single call
to `init_console()`. This allows us to make `WRITER` and its type,
Console, private.

While we are at it, return an error from `init_console()` and handle
it at the caller instead of implicitly panicking.

Signed-off-by: Carlos López <carlos.lopez@suse.com>