platform: abstract PageStateChangeOp
PageStateChangeOp might make sense on multiple platforms, or it can
be adapted in the future. Move it to the platform module and take it
as a parameter for SvsmPlatform::page_state_change(), making calls to
this method more readable.

While we are at it, remove the Psc prefix for the enum variants, as it
is redundant.

Signed-off-by: Carlos López <carlos.lopez@suse.com>