kernel/insn_decode: Add opcode support
Implement the OpCodeDesc structure to represent a decoded opcode, which
includes the instruction type and associated flags to facilitate
upcoming instruction decoding. Currently the opcodes supported are
in/out, wrmsr/rdmsr, cpuid, and rdtsc/rdtscp, which are essential for
the SEV #VC handler. These supported instructions are defined within
the static OpCodeDesc tables. The decode method consults these tables
using the relevant opcode and returns Some(OpCodeDesc) if the opcode is
recognized or None otherwise.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>