kernel/cpu/idt: Implement InsnMachineCtx for X86ExceptionContext
The X86ExceptionContext is used to capure the CPU's state when an
exception is raised. It is utilized by the #VC handler to retrieve the
instruction bytes that triggered the #VC. To enable the insn_decode
crate to decode these instruction bytes according to the current CPU's
state, implement InsnMachineCtx trait for the X86ExceptionContext.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>