kernel/insn_decode: Refine the in and out DecodedInsn
The DecodedInsn for in and out instructions had seperate b/w/l version
denote the length of the input and output bytes. With the ability of
decoding operand length, refine the DecodedInsn for in and out by using
DecodedInsn::In(Operand, Bytes) and DecodedInsn::Out(Operand, Bytes)
respectively. This can simplify the construction of DecodedInsn and io
emulation in the #VC handler.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>