stage1: add trampoline-only compilation
In IGVM mode, the stage2/SVSM binaries no longer need to be loaded from
the stage1 binary. Add trampoline-only compilation to strip the stage1
bootloader down to just a trampoline into stage2.

bin/stage1-trampoline.bin is the output from Makefile that can be used
as the trampoline for stage2.

Signed-off-by: Peter Fang <peter.fang@intel.com>