igvmbuilder: add --tdx-stage1 parameter
The --tdx-stage1 parameter is akin to the --stage2 parameter, except it
is located to end at 4GB. It is meant to be used to bootstrap SVSM from
the reset vector when running as a TD guest. This is needed because TD
guests always start their vCPUs at the reset vector.

As of right now, other platforms can't use the stage1 bootloader in the
IGVM file since it collides with the guest firmware pages at the top of
4GB.

Signed-off-by: Peter Fang <peter.fang@intel.com>