scripts: Fix value for init-flags to ensure measurement is correct
The `init-flags` value in the launch script did not have the 'SVSM' bit
set, which causes VMSA areas to be measured for APs during guest launch.
This does not affect the running of SVSM but does cause the launch
measurement to differ from the value caclculated by igvmmeasure.

This commit sets `init-flags` to 5 to fix this.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>