docs: split FallibleAlloc item into two
The FallibleAlloc item really contains two items. On the one hand,
there is the desire to allow allocations coming from the global
allocator to be fallible. This is blocked by currently unstable
features in the upstream alloc crate.

On the other hand, there is the need for smart pointers that allow
safe allocation from different sources other than the global
allocator. These are higher level abstractions that do not fork or
replace the upstream alloc crate functionality. Once the allocator
infrastructure is stabilized (or forked in), these interfaces can be
implemented on top of it, but there is not a strict dependency.

As such, separate both points for clarity.

Signed-off-by: Carlos López <carlos.lopez@suse.com>