kernel/cpu/vc: Use DecodedInsnCtx for instruction decoding
Refactor the instruction decoding process to utilize DecodedInsnCtx
which provides a context to decode a given set of instruction bytes
along with the associated machine context. The #VC handler now retrieves
the DecodedInsn and its size through the public functions offered by
DecodedInsnCtx. Corresponding updates have been made to the test and
fuzzing code to align with these changes.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>