sev/ghcb: fix GHCB::write_buffer() soundness issuess
Use an usize for the buffer offset, and verify it does not overflow
during bounds checking or when computing the destination pointer.
Also, verify that the destination is properly aligned to T, as an
unaligned write through `ptr::copy_nonoverlapping()` will result in
undefined behavior. On top of that, ensure T is Copy to avoid doing a
bitwise copy of a type that does not support it.

While we are at it, add safety comments.

Signed-off-by: Carlos López <carlos.lopez@suse.com>